# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Multi-Funding Loan
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /MFL
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /orgRegistration:
    x-swagger-router-controller: orgRegistration
    post:
      tags:
      - "Registration"
      summary: "On-boarding organisation"
      description: ""
      operationId: "createOrg"
      produces:
      - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Here you can create Organisations"
          required: true
          schema:
            $ref: "#/definitions/orgRegistration"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/apiResponse"
        400:
          description: Validation error
...
  /login:
    x-swagger-router-controller: LoginController
    post:
      tags:
      - "Registration"
      summary: "On-boarding organisation"
      description: ""
      operationId: "createOrg"
      produces:
      - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Here you can create Organisations"
          required: true
          schema:
            $ref: "#/definitions/orgRegistration"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/apiResponse"
        400:
          description: Validation error
...
definitions:
...
  orgRegistration:
    properties:
      id:
        type: string
      orgId:
        type: string
      orgName:
        type: string
      gstNo:
        type: string
      email:
        type: string
      natureOfBusinessId:
        type: integer
      natureOfBusinessName:
        type: string
      mobile:
        type: "string"
      address:
        type: "string"
      bankName:
        type: string
      bankAccountNo:
        type: string
      ifsc: 
        type: string
  apiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
    example:
      code: 0
      type: "type"
      message: "message"
...